moved volley requests to separate module
