+++ simplified-app-shared/pom.xml
@@ -48,6 +48,13 @@
       <artifactId>simplified-tenprint</artifactId>
       <version>${project.version}</version>
     </dependency>
+
+    <dependency>
+      <groupId>org.nypl.simplified</groupId>
+      <artifactId>simplified-volley</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+
     <dependency>
       <groupId>${project.groupId}</groupId>
       <artifactId>simplified-bugsnag</artifactId>
