+++ simplified-app-shared/src/main/java/org/nypl/simplified/app/reader/ReaderActivity.java
@@ -61,6 +61,7 @@ import org.nypl.simplified.books.core.FeedEntryOPDS;
 import org.nypl.simplified.books.core.LogUtilities;
 import org.nypl.simplified.app.Prefs;
 import org.nypl.simplified.opds.core.OPDSAcquisitionFeedEntry;
+import org.nypl.simplified.volley.NYPLStringRequest;
 import org.readium.sdk.android.Container;
 import org.readium.sdk.android.Package;
 import org.slf4j.Logger;
