+++ simplified-jobqueue/pom.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <parent>
+    <artifactId>simplified</artifactId>
+    <groupId>org.nypl.simplified</groupId>
+    <version>1.1.1</version>
+  </parent>
+  <modelVersion>4.0.0</modelVersion>
+
+  <artifactId>simplified-jobqueue</artifactId>
+
+
+<dependencies>
+  <dependency>
+    <groupId>android</groupId>
+    <artifactId>android</artifactId>
+    <scope>provided</scope>
+  </dependency>
+
+  <dependency>
+    <groupId>android.support</groupId>
+    <artifactId>compatibility-v7-appcompat</artifactId>
+    <version>23.1.1</version>
+  </dependency>
+
+  <dependency>
+    <groupId>android.support</groupId>
+    <artifactId>compatibility-v13</artifactId>
+    <version>23.1.1</version>
+  </dependency>
+
+  <dependency>
+    <groupId>com.birbit</groupId>
+    <artifactId>android-priority-jobqueue</artifactId>
+    <version>2.0.1</version>
+  </dependency>
+
+
+</dependencies>
+</project>
