+++ simplified-jobqueue/src/AndroidManifest.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest
+  xmlns:android="http://schemas.android.com/apk/res/android"
+  package="org.nypl.simplified.jobqueue">
+    <uses-sdk
+      android:minSdkVersion="19"
+      android:targetSdkVersion="21"/>
+
+</manifest>
