+++ simplified-volley/src/main/java/org/nypl/simplified/volley/NYPLJsonObjectRequest.java
@@ -1,8 +1,9 @@
-package org.nypl.simplified.app.reader;
+package org.nypl.simplified.volley;
 
 import com.android.volley.AuthFailureError;
 import com.android.volley.Response;
 import com.android.volley.toolbox.JsonObjectRequest;
+import com.io7m.jnull.Nullable;
 
 import net.iharder.Base64;
 
@@ -22,7 +23,9 @@ import java.util.Map;
 
 public class NYPLJsonObjectRequest extends JsonObjectRequest {
 
-  private final AccountCredentials credentials;
+  private @Nullable AccountCredentials credentials;
+  private @Nullable String username;
+  private @Nullable String password;
 
   /**
    * @param method request method
@@ -77,22 +80,51 @@ public class NYPLJsonObjectRequest extends JsonObjectRequest {
 
   }
 
+  /**
+   * @param method request method
+   * @param url request url
+   * @param in_username basic auth username
+   * @param in_password basic auth password
+   * @param listener response listener
+   * @param error_listener error listener
+   */
+  public NYPLJsonObjectRequest(final int method,
+                               final String url,
+                               final String in_username,
+                               final String in_password,
+                               final JSONObject body,
+                               final Response.Listener<JSONObject>  listener,
+                               final Response.ErrorListener error_listener) {
+    super(method, url, body, listener, error_listener);
+    this.username = in_username;
+    this.password = in_password;
+  }
+
 
   @Override
   public Map<String, String> getHeaders() throws AuthFailureError {
 
     final Map<String, String> params = new HashMap<String, String>();
 
-    final AccountBarcode barcode = this.credentials.getBarcode();
-    final AccountPIN pin = this.credentials.getPin();
+    if (this.credentials != null) {
+      final AccountBarcode barcode = this.credentials.getBarcode();
+      final AccountPIN pin = this.credentials.getPin();
 
-    //add oauth/////
+      //add oauth/////
 
-    final String text = barcode.toString() + ":" + pin.toString();
-    final String encoded =
-      Base64.encodeBytes(text.getBytes(Charset.forName("US-ASCII")));
+      final String text = barcode.toString() + ":" + pin.toString();
+      final String encoded =
+        Base64.encodeBytes(text.getBytes(Charset.forName("US-ASCII")));
+      params.put("Authorization", "Basic " + encoded);
+    }
+    else
+    {
+      final String text = this.username + ":" + this.password;
+      final String encoded =
+        Base64.encodeBytes(text.getBytes(Charset.forName("US-ASCII")));
+      params.put("Authorization", "Basic " + encoded);
 
-    params.put("Authorization", "Basic " + encoded);
+    }
     return params;
   }
