+++ simplified-volley/src/main/java/org/nypl/simplified/volley/NYPLStringRequest.java
@@ -1,4 +1,4 @@
-package org.nypl.simplified.app.reader;
+package org.nypl.simplified.volley;
 
 import com.android.volley.AuthFailureError;
 import com.android.volley.Response;
